requirements.txt
django
django-rest-framework
drf-spectacular[sidecar]

pytest
pytest-django

django-stubs
mypy==1.5.0
black==23.7.0
ruff==0.0.284
